divide by float

+++ app/src/main/java/nl/codingwithlinda/pagekeeper/feature_book_detail/book_detail/presentation/interaction/BookDetailState.kt
@@ -33,7 +33,7 @@ data class BookDetailState(
         }
     }
 
-    private val actualReadingProgress = currentElementId.toFloat() / totalElements
+    private val actualReadingProgress = currentElementId.toFloat() / totalElements.toFloat()
     val readingProgress = if (book?.isFinished == true) 1f else actualReadingProgress
 
 }
